+++ CHANGELOG.md
@@ -74,3 +74,9 @@
 -   STW Worker class.
 -   STW Schematic object.
 -   Getters for heroes , workers , schematics , traps, weapons , different types of heroes in campaign profile.
+
+# 1.1.1
+
+-   An important bug fix in README.md
+-   Made [fortnite_api_io](https://github.com/vanxh/fortnite_api_io) library a seperate package.
+-   Survivors squads getter in campaign profile.